+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/mania/ManiaService.java
@@ -317,6 +317,7 @@ public class ManiaService implements InitializingBean, FrontendStatusChangeListe
 
   /**
    * In case the Studio has been re-installed, the players of the VPin Mania accounts should be restored too.
+   * //TODO
    */
   private void checkPlayerRestoring() {
     List<Account> accounts = maniaClient.getAccountClient().getAccounts();